EC-002: Добавил анимацию перехода на главный экран
