--- app/src/main/java/com/supersonic/evercard/features/root/RootViewModel.kt
@@ -1,4 +1,0 @@
-package com.supersonic.evercard.features.root
-
-class RootViewModel {
-}
