+++ app/src/main/java/com/supersonic/evercard/features/root/ui/MainFragmentViewModel.kt
@@ -1,0 +1,4 @@
+package com.supersonic.evercard.features.root.ui
+
+class MainFragmentViewModel {
+}
