+++ app/src/main/java/com/supersonic/evercard/features/root/ui/RootActivity.kt
@@ -1,0 +1,12 @@
+package com.supersonic.evercard.features.root.ui
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.supersonic.evercard.R
+
+class RootActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_root)
+    }
+}
