--- app/src/main/java/com/supersonic/evercard/features/splash/ui/SplashFragment.kt
@@ -1,4 +1,0 @@
-package com.supersonic.evercard.features.splash.ui
-
-class SplashFragment {
-}
