+++ app/src/main/java/com/supersonic/evercard/features/splash/ui/SplashScreenFragment.kt
@@ -1,0 +1,42 @@
+package com.supersonic.evercard.features.splash.ui
+
+import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
+import android.util.Log
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
+import androidx.navigation.navOptions
+import com.supersonic.evercard.R
+import com.supersonic.evercard.databinding.FragmentSplashScreenBinding
+
+class SplashScreenFragment : Fragment() {
+
+    private var _binding: FragmentSplashScreenBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentSplashScreenBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        Handler(Looper.getMainLooper()).postDelayed({
+            findNavController().navigate(R.id.action_splashFragment_to_mainFragment)
+        }, 2000)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
